Sandbox web languages using iframes (#268)
Fixes #239 

Load web libraries in separate pages and embed them using iframes. This
way, these libraries do not pollute the global Window from the session
page and there are no collisions between the globals each library
define.

Incidentally, this also fixes #193: Because the web language libraries
are only loaded inside the iframe and thus in a separate JS context,
when unloading the Iframe DOM element, sound and visuals stop.

Strudel also adds some extensions to the Codemirror editor and has an
animation frame to update decorators. For some reason, these updates
need to run in the same context as the editor, and can't be modified
from within the Iframe page.